+++ dfols/tests/test_trust_region.py
@@ -27,7 +27,7 @@ from math import sqrt
 import numpy as np
 import unittest
 
-from dfols.trust_region import trsbox, trsbox_geometry
+from dfols.trust_region import ctrsbox, ctrsbox_geometry, trsbox, trsbox_geometry
 from dfols.util import model_value
 
 
@@ -69,6 +69,12 @@ def cauchy_pt_box(g, H, delta, lower, upper):
     if crvmin < 0.0:
         crvmin = -1.0
     return s, red, crvmin
+
+def p_box(x,l,u):
+    return np.minimum(np.maximum(x,l), u)
+
+def p_ball(x,c,r):
+    return c + (r/np.max([np.linalg.norm(x-c),r]))*(x-c)
 
 
 class TestUncInternal(unittest.TestCase):
@@ -340,3 +346,135 @@ class TestGeom2WithAlmostZeros2(unittest.TestCase):
         # print(x)
         # print(xtrue)
         self.assertTrue(np.max(np.abs(x - xtrue)) < 1e-10, 'Wrong step')
+
+
+# DFO-LS with arbitrary constraints
+class TestGeom3CDFO(unittest.TestCase):
+    def runTest(self):
+        xbase = np.array([0.0, 0.0]) + 1
+        g = np.array([1.0, -1.0])
+        a = np.array([-2.0, -2.0]) + 1
+        b = np.array([1.0, 2.0]) + 1
+        proj = lambda x: p_box(x,a,b)
+        delta = 5.0
+        c = 3.0  # may want to max instead
+        x = ctrsbox_geometry(xbase, c, g, [proj], delta)
+        xtrue = np.array([1.0, -2.0])
+        self.assertTrue(np.max(np.abs(x - xtrue)) < 1e-10, 'Wrong step')
+
+class TestUncInternalCDFO(unittest.TestCase):
+    def runTest(self):
+        n = 3
+        g = np.array([1.0, 0.0, 1.0])
+        H = np.array([[1.0, 0.0, 0.0], [0.0, 2.0, 0.0], [0.0, 0.0, 2.0]])
+        Delta = 2.0
+        xopt = np.ones((n,))  # trying nonzero (since bounds inactive)
+        d, gnew, _crvmin = ctrsbox(xopt, g, H, [], Delta)
+        true_d = np.array([-1.0, 0.0, -0.5])
+        est_min = model_value(g, H, d)
+        true_min = model_value(g, H, true_d)
+        s_cauchy, red_cauchy, _crvmin_cauchy = cauchy_pt(g, H, Delta)
+        self.assertTrue(est_min <= red_cauchy, 'Cauchy reduction not achieved')
+        self.assertTrue(np.all(gnew == g + H.dot(d)), 'Wrong gnew')
+
+class TestUncBdryCDFO(unittest.TestCase):
+    def runTest(self):
+        n = 3
+        g = np.array([1.0, 0.0, 1.0])
+        H = np.array([[1.0, 0.0, 0.0], [0.0, 2.0, 0.0], [0.0, 0.0, 2.0]])
+        Delta = 5.0 / 12.0
+        xopt = np.zeros((n,))
+        d, gnew, _crvmin = ctrsbox(xopt, g, H, [], Delta)
+        true_d = np.array([-1.0 / 3.0, 0.0, -0.25])
+        est_min = model_value(g, H, d)
+        true_min = model_value(g, H, true_d)
+        s_cauchy, red_cauchy, _crvmin_cauchy = cauchy_pt(g, H, Delta)
+        self.assertTrue(est_min <= red_cauchy, 'Cauchy reduction not achieved')
+        self.assertTrue(np.allclose(gnew, g + H.dot(d)), 'Wrong gnew')
+
+
+class TestUncHardCDFO(unittest.TestCase):
+    def runTest(self):
+        n = 3
+        g = np.array([0.0, 0.0, 1.0])
+        H = np.array([[-2.0, 0.0, 0.0], [0.0, -1.0, 0.0], [0.0, 0.0, -1.0]])
+        Delta = sqrt(2.0)
+        xopt = np.zeros((n,))
+        d, gnew, _crvmin = ctrsbox(xopt, g, H, [], Delta)
+        true_d = np.array([1.0, 0.0, -1.0])  # non-unique solution
+        est_min = model_value(g, H, d)
+        true_min = model_value(g, H, true_d)
+        s_cauchy, red_cauchy, _crvmin_cauchy = cauchy_pt(g, H, Delta)
+        self.assertTrue(est_min <= red_cauchy, 'Cauchy reduction not achieved')
+        self.assertTrue(np.allclose(gnew, g + H.dot(d)), 'Wrong gnew')
+
+class TestConInternalCDFO(unittest.TestCase):
+    def runTest(self):
+        n = 3
+        g = np.array([1.0, 0.0, 1.0])
+        H = np.array([[1.0, 0.0, 0.0], [0.0, 2.0, 0.0], [0.0, 0.0, 2.0]])
+        Delta = 2.0
+        xopt = np.ones((n,))  # trying nonzero (since bounds inactive)
+        sl = xopt + np.array([-0.5, -10.0, -10.0])
+        su = xopt + np.array([10.0, 10.0, 10.0])
+        proj = lambda x: p_box(x,sl,su)
+        d, gnew, _crvmin = ctrsbox(xopt, g, H, [proj], Delta)
+        true_d = np.array([-1.0, 0.0, -0.5])
+        est_min = model_value(g, H, d)
+        true_min = model_value(g, H, true_d)
+        s_cauchy, red_cauchy, _crvmin_cauchy = cauchy_pt_box(g, H, Delta, sl-xopt, su-xopt)
+        self.assertTrue(est_min <= red_cauchy, 'Cauchy reduction not achieved')
+        self.assertTrue(np.all(gnew == g + H.dot(d)), 'Wrong gnew')
+
+class TestConBdryCDFO(unittest.TestCase):
+    def runTest(self):
+        n = 3
+        g = np.array([1.0, 0.0, 1.0])
+        H = np.array([[1.0, 0.0, 0.0], [0.0, 2.0, 0.0], [0.0, 0.0, 2.0]])
+        Delta = 5.0 / 12.0
+        xopt = np.zeros((n,))
+        sl = xopt + np.array([-0.3, -0.01, -0.1])
+        su = xopt + np.array([10.0, 1.0, 10.0])
+        proj = lambda x: p_box(x,sl,su)
+        d, gnew, _crvmin = ctrsbox(xopt, g, H, [proj], Delta)
+        true_d = np.array([-1.0 / 3.0, 0.0, -0.25])
+        est_min = model_value(g, H, d)
+        true_min = model_value(g, H, true_d)
+        s_cauchy, red_cauchy, _crvmin_cauchy = cauchy_pt_box(g, H, Delta, sl - xopt, su - xopt)
+        self.assertTrue(est_min <= red_cauchy, 'Cauchy reduction not achieved')
+        self.assertTrue(np.max(np.abs(gnew - g - H.dot(d))) < 1e-10, 'Wrong gnew')
+
+class TestBoxBallInternalCDFO(unittest.TestCase):
+    def runTest(self):
+        n = 3
+        g = np.array([1.0, 0.0, 1.0])
+        H = np.array([[1.0, 0.0, 0.0], [0.0, 2.0, 0.0], [0.0, 0.0, 2.0]])
+        Delta = 2.0
+        xopt = np.ones((n,))  # trying nonzero (since bounds inactive)
+        sl = xopt + np.array([-0.5, -10.0, -10.0])
+        su = xopt + np.array([10.0, 10.0, 10.0])
+        boxproj = lambda x: p_box(x,sl,su)
+        ballproj = lambda x: p_ball(x,xopt,5)
+        d, gnew, _crvmin = ctrsbox(xopt, g, H, [boxproj,ballproj], Delta)
+        true_d = np.array([-0.5, 0.0, -0.5])
+        est_min = model_value(g, H, d)
+        true_min = model_value(g, H, true_d)
+        self.assertTrue(est_min <= true_min + 1e-3, 'Sufficient decrease not achieved')
+
+
+class TestBoxBallBdryCDFO(unittest.TestCase):
+    def runTest(self):
+        n = 3
+        g = np.array([1.0, 0.0, 1.0])
+        H = np.array([[1.0, 0.0, 0.0], [0.0, 2.0, 0.0], [0.0, 0.0, 2.0]])
+        Delta = 5.0 / 12.0
+        xopt = np.zeros((n,))
+        sl = xopt + np.array([-0.3, -0.01, -0.1])
+        su = xopt + np.array([10.0, 1.0, 10.0])
+        boxproj = lambda x: p_box(x,sl,su)
+        ballproj = lambda x: p_ball(x,xopt,0.25)
+        d, gnew, _crvmin = ctrsbox(xopt, g, H, [boxproj,ballproj], Delta)
+        true_d = np.array([-0.22913085, 0.0, -0.09999527])
+        est_min = model_value(g, H, d)
+        true_min = model_value(g, H, true_d)
+        self.assertTrue(est_min <= true_min + 1e-3, 'Sufficient decrease not achieved')